feat: estrutura inicial clean architecture
