+++ src/main/java/br/fiap/projeto/contexto/pedido/infrastructure/configuration/BeanPedidoConfiguration.java
@@ -3,9 +3,6 @@ package br.fiap.projeto.contexto.pedido.infrastructure.configuration;
 import br.fiap.projeto.contexto.pedido.domain.port.repository.PedidoRepositoryPort;
 import br.fiap.projeto.contexto.pedido.domain.port.service.PedidoService;
 import br.fiap.projeto.contexto.pedido.domain.service.DomainPedidoService;
-import br.fiap.projeto.contexto.produto.domain.port.repository.ProdutoRepositoryPort;
-import br.fiap.projeto.contexto.produto.domain.port.service.ProdutoServicePort;
-import br.fiap.projeto.contexto.produto.domain.service.DomainProdutoService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
