+++ src/main/java/br/fiap/projeto/contexto/produto/adapter/controller/ProdutoAdapterController.java
@@ -1,0 +1,61 @@
+package br.fiap.projeto.contexto.produto.adapter.controller;
+
+import br.fiap.projeto.contexto.produto.adapter.controller.rest.request.ProdutoDTORequest;
+import br.fiap.projeto.contexto.produto.adapter.controller.rest.response.ProdutoDTOResponse;
+import br.fiap.projeto.contexto.produto.adapter.interfaces.IProdutoAdapterController;
+import br.fiap.projeto.contexto.produto.entity.Produto;
+import br.fiap.projeto.contexto.produto.entity.enums.CategoriaProduto;
+import br.fiap.projeto.contexto.produto.usecase.exception.EntradaInvalidaException;
+import br.fiap.projeto.contexto.produto.usecase.exception.ProdutoNaoEncontradoException;
+import br.fiap.projeto.contexto.produto.usecase.port.service.IProdutoUseCase;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class ProdutoAdapterController implements IProdutoAdapterController {
+
+    private final IProdutoUseCase produtoUseCase;
+
+    public ProdutoAdapterController(IProdutoUseCase produtoUseCase) {
+        this.produtoUseCase = produtoUseCase;
+    }
+
+    @Override
+    public List<ProdutoDTOResponse> getProdutos() {
+        List<Produto> produtos = this.produtoUseCase.buscaTodos();
+        return produtos.stream().map(ProdutoDTOResponse::newInstanceByProduto).collect(Collectors.toList());
+    }
+
+    @Override
+    public ProdutoDTOResponse getProduto(String codigo) throws ProdutoNaoEncontradoException {
+        Produto produto = this.produtoUseCase.buscaProduto(codigo);
+        return ProdutoDTOResponse.newInstanceByProduto(produto);
+    }
+
+    @Override
+    public List<ProdutoDTOResponse> getProdutosPorCategoria(CategoriaProduto categoria) {
+        List<Produto> produtos = this.produtoUseCase.buscaProdutosPorCategoria(categoria);
+        return produtos.stream().map(ProdutoDTOResponse::newInstanceByProduto).collect(Collectors.toList());
+    }
+
+    @Override
+    public List<String> getCategoriasDeProdutos() {
+        return this.produtoUseCase.getCategoriasDeProdutos();
+    }
+
+    @Override
+    public ProdutoDTOResponse criaProduto(ProdutoDTORequest produtoDTORequest) throws EntradaInvalidaException {
+        Produto produto = this.produtoUseCase.criaProduto(produtoDTORequest.toProduto());
+        return ProdutoDTOResponse.newInstanceByProduto(produto);
+    }
+
+    @Override
+    public void removeProduto(String codigo) throws ProdutoNaoEncontradoException {
+        this.produtoUseCase.removeProduto(codigo);
+    }
+
+    @Override
+    public void atualizaProduto(String codigo, ProdutoDTORequest produtoDTO) throws ProdutoNaoEncontradoException {
+        this.produtoUseCase.atualizaProduto(codigo, produtoDTO.toProduto());
+    }
+}
