--- src/main/java/br/fiap/projeto/contexto/produto/domain/enums/CategoriaProduto.java
@@ -1,6 +1,0 @@
-package br.fiap.projeto.contexto.produto.domain.enums;
-
-public enum CategoriaProduto {
-    LANCHE, ACOMPANHAMENTO, BEBIDA, SOBREMESA;
-}
-
