--- src/main/java/br/fiap/projeto/contexto/produto/domain/exception/EntradaInvalidaException.java
@@ -1,14 +1,0 @@
-package br.fiap.projeto.contexto.produto.domain.exception;
-
-public class EntradaInvalidaException extends BaseException {
-    private final static String DEFAULT_MESSAGE = "Entrada inválida!";
-    private final static int CODE = 3003;
-
-    public EntradaInvalidaException() {
-        super(CODE, DEFAULT_MESSAGE);
-    }
-
-    public EntradaInvalidaException(String message) {
-        super(CODE, message);
-    }
-}
