+++ src/main/java/br/fiap/projeto/contexto/produto/entity/enums/CategoriaProduto.java
@@ -1,0 +1,6 @@
+package br.fiap.projeto.contexto.produto.entity.enums;
+
+public enum CategoriaProduto {
+    LANCHE, ACOMPANHAMENTO, BEBIDA, SOBREMESA;
+}
+
